CCFDAPI/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
joblib
